Metodos da categoria corrigidos

+++ Servidor/src/main/java/com/example/Servidor/controlers/CategoryControler.java
@@ -3,6 +3,9 @@ package com.example.Servidor.controlers;
 
 import com.example.Servidor.entities.Category;
 import com.example.Servidor.services.CategoryService;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -10,29 +13,54 @@ import java.util.List;
 @RestController
 @RequestMapping("/categorias")
 public class CategoryControler {
-    private final CategoryService categoryService;
 
-    public CategoryControler (CategoryService categoryService){
-        this.categoryService = categoryService;
+    @Autowired
+    private CategoryService categoryService;
+
+    public CategoryControler (){
     }
 
-    @GetMapping
-    public List<Category> list(){
-       return categoryService.listAll();
-    }
 
-    @GetMapping("/usuarios/userId")
-    public List<Category> listForUser(@PathVariable int userId){
-        return categoryService.listForUser(userId);
+    @GetMapping("/{userId}")
+    public ResponseEntity<List<Category>> listForUser(@PathVariable int userId){
+        if (categoryService.listForUser(userId).isEmpty()){
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.ok().body(categoryService.listForUser(userId));
     }
 
     @PostMapping
-    public Category createCategory(@RequestBody Category category){
-        return categoryService.saveCategory(category);
+    public ResponseEntity<Category> createCategory(@RequestBody Category category){
+        if(categoryService.createCategory(category) == null){
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.ok().body(category);
+        
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Category> updateCategory(
+        @PathVariable Integer id, 
+        @RequestBody Category category){
+
+        if(categoryService.updateCategory(id, category) == null){
+            return ResponseEntity.notFound().build();
+        }
+        
+        return ResponseEntity.ok().body(category);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCategory(@PathVariable int id){
+    public ResponseEntity<Void> deleteCategory(@PathVariable int id){
+        if(!categoryService.existsById(id)){
+            return ResponseEntity.notFound().build();
+        }
+
         categoryService.deleteCategory(id);
+        if(!categoryService.existsById(id)){
+            return ResponseEntity.noContent().build();
+        }
+
+        return ResponseEntity.internalServerError().build();
     }
 }
